+++ Cargo.toml
@@ -78,6 +78,7 @@ members = [
     "common/nym-kkt-ciphersuite",
     "common/nym-kkt-context",
     "common/nym-lp",
+    "common/nym-lp-data",
     "common/nym-metrics",
     "common/nym_offline_compact_ecash",
     "common/nymnoise",
@@ -459,6 +460,7 @@ nym-id = { version = "1.21.0", path = "common/nym-id" }
 nym-ip-packet-client = { version = "1.21.0", path = "nym-ip-packet-client" }
 nym-ip-packet-requests = { version = "1.21.0", path = "common/ip-packet-requests" }
 nym-lp = { version = "1.21.0", path = "common/nym-lp" }
+nym-lp-data = { version = "1.21.0", path = "common/nym-lp-data" }
 nym-kkt = { version = "1.21.0", path = "common/nym-kkt" }
 nym-kkt-ciphersuite = { version = "1.21.0", path = "common/nym-kkt-ciphersuite" }
 nym-kkt-context = { version = "1.21.0", path = "common/nym-kkt-context" }
