+++ common/nym-lp-data/src/lib.rs
@@ -1,0 +1,4 @@
+// Copyright 2026 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod packet;
