+++ common/nym-lp-data/src/packet/error.rs
