+++ nym-node/src/node/lp/control/ingress/client_handler.rs
@@ -6,13 +6,15 @@ use crate::node::lp::control::{LP_DURATION_BUCKETS, LpConnectionStats};
 use crate::node::lp::error::LpHandlerError;
 use crate::node::lp::state::SharedLpClientControlState;
 use dashmap::mapref::one::RefMut;
-use nym_lp::packet::frame::LpFrameKind;
-use nym_lp::packet::{EncryptedLpPacket, ForwardPacketData, LpFrame};
-use nym_lp::peer_config::LpReceiverIndex;
+use nym_lp::LpTransportSession;
 use nym_lp::session::{LpAction, LpInput};
 use nym_lp::transport::LpHandshakeChannel;
 use nym_lp::transport::traits::LpTransportChannel;
-use nym_lp::{LpTransportSession, packet::frame::ExpectedResponseSize};
+use nym_lp_data::packet::frame::LpFrameKind;
+use nym_lp_data::packet::header::LpReceiverIndex;
+use nym_lp_data::packet::{
+    EncryptedLpPacket, ForwardPacketData, LpFrame, frame::ExpectedResponseSize,
+};
 use nym_metrics::{add_histogram_obs, inc};
 use nym_node_metrics::NymNodeMetrics;
 use nym_registration_common::{LpRegistrationRequest, RegistrationStatus};
